Move_optimal_to_b

+++ libft/ft_atoi.c
@@ -6,17 +6,17 @@
 /*   By: arakotoa <arakotoa@student.42antananari    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/12 09:38:26 by arakotoa          #+#    #+#             */
-/*   Updated: 2024/03/19 14:54:18 by arakotoa         ###   ########.fr       */
+/*   Updated: 2024/11/25 17:25:02 by arakotoa         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_atoi(const char *nptr)
+long	ft_atoi(const char *nptr)
 {
-	int	i;
-	int	sign;
-	int	result;
+	int		i;
+	int		sign;
+	long	result;
 
 	i = 0;
 	sign = 1;
